# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Arabic RTL order, inventory, delivery, and financial management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadRequest"
      responses:
        "200":
          description: Upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
  /storage/objects/{objectPath}:
    get:
      operationId: getStoredObject
      tags: [storage]
      summary: Serve a stored private object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stored file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
components:
...
  schemas:
...
    UploadRequest:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        size: { type: integer, minimum: 1 }
        contentType: { type: string, minLength: 1 }
      required: [name, size, contentType]
    UploadResponse:
      type: object
      properties:
        uploadURL: { type: string, format: uri }
        objectPath: { type: string }
      required: [uploadURL, objectPath]