# lib/api-zod/src/generated/types/uploadRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Arabic RTL order, inventory, delivery, and financial management API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}